% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipainfo.R
\name{ipainfo}
\alias{ipainfo}
\title{IPA Information Database}
\usage{
ipainfo(wanted)
}
\value{
A list with the following components:

\item{symbol}{IPA Unicode hexadecimal representations of 30 vowel symbols.}
\item{description}{Phonetic descriptions (e.g., "close front unrounded").}
\item{num}{Numeric identifiers (1-30) for each vowel.}
\item{chooseplot}{Data frame with vowel chart coordinates: height, frontness,
and rounded (boolean).}
\item{sampa}{X-SAMPA representations of the IPA vowels.}
}
\description{
Retrieve information about IPA vowel symbols.
}
\details{
This function returns a list containing comprehensive information about IPA
(International Phonetic Alphabet) vowel symbols, including their Unicode
representations, phonetic descriptions, and coordinates for visualization on
a vowel chart.
}
\references{
https://www.internationalphoneticassociation.org/
}
\author{
Santiago Barreda <sbarreda@ucdavis.edu>
}
\keyword{internal}
