% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iplot.R
\name{iplot}
\alias{iplot}
\title{Interaction Plot}
\usage{
iplot(
  y,
  x,
  lines,
  group = NULL,
  ylab = "",
  xlabels = c("", ""),
  bars = "se",
  add = FALSE,
  cols = c(1, 1)
)
}
\arguments{
\item{y}{A numeric vector of response values.}

\item{x}{A factor or grouping variable for the x-axis.}

\item{lines}{A factor indicating different lines to be plotted.}

\item{group}{An optional grouping variable for additional stratification.}

\item{ylab}{Label for the y-axis.}

\item{xlabels}{Labels for the x-axis positions.}

\item{bars}{Type of error bars: 'se' for standard error or 'sd' for
standard deviation.}

\item{add}{If TRUE, lines are added to an existing plot. If FALSE, a new
plot is created.}

\item{cols}{A vector of colors for the plotted lines.}
}
\description{
Create an interaction plot with error bars.
}
\details{
This function generates an interaction plot showing the relationship between
two factors with optional error bars (standard error or standard deviation).
It can accommodate grouping variables and allows customization of colors and
labels.
}
\author{
Santiago Barreda <sbarreda@ucdavis.edu>
}
