% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p73}
\alias{p73}
\title{Pols et al. (1973) Dutch Vowel Data}
\format{
A data frame with the following 9 variables:
\itemize{
  \item sex A factor indicating speaker sex.
  \item vowel The vowel category in x-sampa.
  \item f1 A numeric vector indicating the vowel F1 in Hz.
  \item f2 A numeric vector indicating the vowel F2 in Hz.
  \item f3 A numeric vector indicating the vowel F3 in Hz.
}
}
\description{
Formant frequency information for vowels averaged across 24 male speakers.
}
\examples{

data(p73)
vplot (p73$f1, p73$f2, p73$vowel, logaxes = TRUE, xsampa = FALSE)

}
\references{
Pols, L. C. W., Tromp, H. R. C., & Plomp, R. (1973). Frequency
analysis of Dutch vowels from 50 male speakers. Journal of the Acoustical
Society of America, 53. 1093-1101.
}
