% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{phonTools-package}
\alias{phonTools}
\alias{phonTools-package}
\title{phonTools: Phonetic and Acoustic Analysis Tools}
\description{
A comprehensive R package for phonetic and acoustic analysis with functions
for signal processing, formant extraction, vowel analysis, and visualization.
}
\author{
\strong{Maintainer}: Santiago Barreda \email{sbarreda@ucdavis.edu}

}
\keyword{internal}
