% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sound.R
\name{plot.sound}
\alias{plot.sound}
\title{Plot Sound Objects}
\usage{
\method{plot}{sound}(x, ...)
}
\arguments{
\item{x}{A sound object to be plotted.}

\item{...}{Additional graphical parameters passed to the plot function.}
}
\value{
Invisibly returns NULL. Used for side effects (plotting).
}
\description{
Plot the waveform of a sound object.
}
\details{
Creates a time-domain plot of a sound object's waveform. If the sound is
stored as a time series (ts) object, R's built-in plotting is used.
Otherwise, time is computed from the sampling frequency.
}
\examples{
data(sound)
plot(sound)

}
\author{
Santiago Barreda <sbarreda@ucdavis.edu>
}
