% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.spectrogram.R
\name{plot.spectrogram}
\alias{plot.spectrogram}
\title{Plot Spectrograms}
\usage{
\method{plot}{spectrogram}(x, y, ylim, xlim, quality = FALSE, ...)
}
\arguments{
\item{x}{A spectrogram object to be plotted.}

\item{y}{Unused (for compatibility with generic plot method).}

\item{ylim}{Y-axis limits (frequency range).}

\item{xlim}{X-axis limits (time range).}

\item{quality}{If TRUE, higher quality output is produced (slower).}

\item{...}{Additional graphical parameters passed to image function.}
}
\value{
Invisibly returns NULL. Used for side effects (plotting).
}
\description{
Plot a spectrogram object as a heatmap.
}
\details{
Creates a color-coded time-frequency representation of a spectrogram object.
Supports customizable color schemes, dynamic range, and frequency/time limits.
}
\examples{
data(sound)
sp <- spectrogram(sound)
plot(sp)

}
\author{
Santiago Barreda <sbarreda@ucdavis.edu>
}
