% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.sound.R
\name{print.sound}
\alias{print.sound}
\title{Print Sound Objects}
\usage{
\method{print}{sound}(x, ...)
}
\arguments{
\item{x}{A sound object to be printed.}

\item{...}{Unused, for compatibility with generic print method.}
}
\value{
Invisibly returns x. Used for side effects (printing).
}
\description{
Display information about a sound object.
}
\details{
Prints summary information about a sound object including the filename,
sampling frequency, duration, and number of samples.
}
\examples{
data(sound)
print(sound)

}
\author{
Santiago Barreda <sbarreda@ucdavis.edu>
}
