% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.spectrogram.R
\name{print.spectrogram}
\alias{print.spectrogram}
\title{Print Spectrogram Objects}
\usage{
\method{print}{spectrogram}(x, ...)
}
\arguments{
\item{x}{A spectrogram object to be printed.}

\item{...}{Unused, for compatibility with generic print method.}
}
\value{
Invisibly returns x. Used for side effects (printing).
}
\description{
Display information about a spectrogram object.
}
\details{
Prints a brief message about a spectrogram object and indicates that it
should be plotted using plot().
}
\examples{
data(sound)
sp <- spectrogram(sound)
print(sp)

}
\author{
Santiago Barreda <sbarreda@ucdavis.edu>
}
