% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvtnorm.R
\name{rmvtnorm}
\alias{rmvtnorm}
\title{Random multivariate normal variables}
\usage{
rmvtnorm(n = 1, k = 2, means = rep(0, k), sigma = diag(k))
}
\arguments{
\item{n}{The number of vectors to be drawn.}

\item{k}{The dimension of the vectors to be drawn.}

\item{means}{A vector of means, one for each dimension.}

\item{sigma}{The covariance matrix of the distribution. If a number between
0 and 1 is provided, this is assumed to be the correlation parameter for a
bivariate standard normal distribution.}
}
\value{
A matrix with rows equal to n and columns equal to k, where each row
indicates a single observation, and each column represents a different
dimension.
}
\description{
Draw vectors from a multivariate normal distribution.
}
\details{
If means and sigma are not specified, a standard normal distribution is
assumed along every dimensions, and dimensions are assumed to be
uncorrelated. If the number of dimensions is not specified, a bivariate
distribution is assumed.
}
\examples{

## Examples of draws from different bivariate normal distributions
## and standard deviation ellipses drawn to fit them.
par (mfrow = c(2,2))
draws = rmvtnorm (n = 1000, k = 2, sigma = .3)
plot (draws)
sdellipse (draws, stdev = 3, lwd = 3, col = 2)

draws = rmvtnorm (n = 1000, k = 2, sigma = -.3)
plot (draws)
sdellipse (draws, stdev = 3, lwd = 3, col = 2)

draws = rmvtnorm (n = 1000, k = 2, sigma = -.7)
plot (draws)
sdellipse (draws, stdev = 3, lwd = 3, col = 2)

draws = rmvtnorm (n = 1000, k = 2, sigma = .7)
plot (draws)
sdellipse (draws, stdev = 3, lwd = 3, col = 2)

}
\author{
Santiago Barreda <sbarreda@ucdavis.edu>
}
