% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sound}
\alias{sound}
\title{Sound object}
\format{
This sound object has the same properties as all sound objects.
These may be inspected by using the print() function.
}
\description{
An example of a sound object, the phrase 'This is a spectrogram', produced
by the author of this package. This sound object may be inspected with
spectrogram() and plotted with plot(). Sections of the individual samples
representing the waveform (found in the "sound" element of a sound object)
can be passed to spectralslice() to see a spectral slice. Several functions
included in this package may also be used to modify or manipulate 'sound'
objects.
}
\examples{


## uncomment and run

#data (sound)
#par (mfrow = c(1,2), mar = c(4,4,1,1))

#plot (sound)
#spectrogram (sound, dynamicrange = 50, maxfreq = 7000)

}
