% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{t07}
\alias{t07}
\title{Thomson (2007) Vowel Data}
\format{
A data frame with 1520 observations on the following 9 variables:
\itemize{
  \item gender A factor indicating speaker sex.
  \item vowel The vowel category in x-sampa.
  \item f1 A numeric vector indicating the vowel F1 in Hz.
  \item f2 A numeric vector indicating the vowel F2 in Hz.
  \item f3 A numeric vector indicating the vowel F3 in Hz.
  \item f4 A numeric vector indicating the vowel F4 in Hz.
  }
}
\description{
Mean formant frequency information for the vowels of male and female
speakers from Edmonton English.
}
\examples{

data(t07)
vplot (t07$f1, t07$f2, t07$vowel, logaxes = TRUE, meansOnly = TRUE, 
xsampa = FALSE)

}
\references{
Thomson, R. (2007). Modeling L1/L2 interactions in the
perception and production of English vowels by Mandarin L1 speakers: A
training study. PhD dissertation, University of Alberta.
}
