% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{voweldata}
\alias{voweldata}
\title{Information about Vowel Data Sets}
\format{
A data frame with the following information: 
\itemize{
  \item name The name of the data set.
  \item language The language represented in the data set.
  \item nspeakers The number of speakers
  \item sexes The sexes of the speakers, m (male) and female (f).
  \item nformants The number of formants represented, between 2 and 4.
  \item individual If 'yes', individual speaker measurements are given, if 'no', measurements are averaged within gender.
}
}
\description{
Contains a brief description and listing of all vowel data sets included in
this package.
}
\examples{

data(voweldata)

}
\references{
At the moment 9 vowel data sets are included in this package.
See individual data set pages for citation information.
}
