% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{y96}
\alias{y96}
\title{Yang (1996) Korean Vowel Data}
\format{
A data frame with the following columns: 
\itemize{
  \item sex A factor indicating speaker sex.
  \item vowel The vowel category in x-sampa.
  \item f1 A numeric vector indicating the vowel F1 in Hz.
  \item f2 A numeric vector indicating the vowel F2 in Hz.
  \item f3 A numeric vector indicating the vowel F3 in Hz.
  }
}
\description{
Formant frequency information for vowels averaged across 60 male and female
speakers.
}
\examples{

data(y96)
vplot (y96$f1, y96$f2, y96$vowel, logaxes = TRUE, xsampa = FALSE)

}
\references{
Yang, B. (1996). A comparative study of American English and
Korean vowels produced by male and female speakers. Journal of Phonetics,
24. 245-261.
}
