% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{f73}
\alias{f73}
\title{Fant (1973) Swedish Vowel Data}
\format{
A data frame with the following columns: 
\itemize{
  \item sex A factor indicating speaker sex.
  \item vowel The vowel category in x-sampa.
  \item f1 A numeric vector indicating the vowel F1 in Hz.
  \item f2 A numeric vector indicating the vowel F2 in Hz.
  \item f3 A numeric vector indicating the vowel F3 in Hz.
  \item f4 A numeric vector indicating the vowel F4 in Hz.
  }
}
\description{
Formant frequency information for vowels averaged across 50 male speakers.
}
\examples{

data(f73)
vowelplot (f73$f1, f73$f2, f73$vowel, logaxes = 'xy', xsampa = TRUE)

}
\references{
Fant, G. (1973). Speech sounds and features. Cambridge, MA: MIT
Press.
}
