% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadtable.R
\name{loadtable}
\alias{loadtable}
\title{Load Table}
\usage{
loadtable(...)
}
\arguments{
\item{...}{Arguments are passed to the internal call of read.table().}
}
\value{
Returns the output of read.table().
}
\description{
Load text table data quickly.
}
\details{
This function is a wrapper for read.table() for those times when you just
don't feel like typing a filename. The function opens up a file selection
dialog box allowing the user to select the file containing the data.
}
\examples{

## uncomment and run
# data = loadtable ()
# head (data)

}
\author{
Santiago Barreda <sbarreda@ucdavis.edu>
}
