% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntypes.R
\name{ntypes}
\alias{ntypes}
\title{Number of Unique Elements}
\usage{
ntypes(vector)
}
\arguments{
\item{vector}{The vector of interest.}
}
\description{
Find the number of unique elements in a vector.
}
\details{
A simple function that converts a vector to a factor, and finds the number
of levels. This provides the number of unique elements in a vector,
something I find I frequently need.
}
\examples{

data (pb52)
## the number of unique vowel categories.
ntypes (pb52$vowel)

}
\author{
Santiago Barreda <sbarreda@ucdavis.edu>
}
