% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pb52}
\alias{pb52}
\title{Peterson & Barney (1952) Vowel Data}
\format{
A data frame with 1520 observations on the following 9 variables:
\itemize{
  \item sex A factor indicating speaker sex.
  \item vowel The vowel category in x-sampa.
  \item f1 A numeric vector indicating the vowel F1 in Hz.
  \item f2 A numeric vector indicating the vowel F2 in Hz.
  \item f3 A numeric vector indicating the vowel F3 in Hz.
  \item f0 A numeric vector indicating the vowel f0 in Hz.
  \item dur A numeric vector indicating the vowel duration in ms.
  \item repetition A numeric vector indicating the repetition number.
  \item speaker A numeric vector indicating a uniquely identifying speaker number.
  \item type A factor with levels c m and w representing speaker type: child, man and woman.
  }
}
\source{
The data was created from tables provided within Praat:

http://www.fon.hum.uva.nl/praat/
}
\description{
Formant frequency and f0 information for vowels collected from 76 speakers
in the Peterson & Barney (1952) data.
}
\examples{

data(pb52)
vowelplot (pb52$f1, pb52$f2, pb52$vowel, logaxes = 'xy', ellipses = TRUE, 
xsampa = TRUE)

}
\references{
Peterson, G.E. & Barney (1952). Control methods used in a study
of the vowels. Journal of the Acoustical Society of America 24: 175-184.

Boersma, Paul & Weenink, David (2012). Praat: doing phonetics by computer
[Computer program]. Version 5.3.19, retrieved 24 June 2012 from
http://www.praat.org/
}
