% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.fraction.R
\name{reduce.fraction}
\alias{reduce.fraction}
\title{Reduce Fractions}
\usage{
reduce.fraction(ratio)
}
\arguments{
\item{ratio}{A vector with two integers. The first element is the numerator
of a ratio, and the second element is the denominator.}
}
\value{
A vector containing the elements of the reduced fraction. The first
element is the numerator of a ratio, and the second element is the
denominator.
}
\description{
Reduce fractions to lowest terms using Euclid's Algorithm.
}
\examples{

## an easy one
reduce.fraction (c(100, 200))

## irreducible
reduce.fraction (c(140, 201))

## a hard one
reduce.fraction (c(140, 203))

}
\author{
Santiago Barreda <sbarreda@ucdavis.edu>
}
